Add files via upload
This is a money manager for children.  IT uses local storage within the browser to store the variables.  It is not secure, but is not intended to be.  It has a login page, and a money manager.  You can populate an array and assign the serial numbers to real monopoly like money, then the kids can redeem them with the program.  It's not perfect but it's neat.